commit2

+++ src/pedido/Item.java
@@ -2,17 +2,17 @@ package pedido;
 
 public class Item {
 
-    private Long id;
+    private Integer id;
     private String nome;
     private Double preco;
 
-    public Item(Long id, String nome, Double preco) {
+    public Item(Integer id, String nome, Double preco) {
         this.id = id;
         this.nome = nome;
         this.preco = preco;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
